requirements.txt
click
pyyaml
cf
netcdf4

